Fix Codex credit meter fallback
